requirements.txt
PySide6~=6.6.3.1
mouse~=0.7.1
requests~=2.31.0
shortuuid~=1.0.13
pillow~=10.3.0
